Refactor instance sdk helper function to container.go file
